git: Add base branch support to create_branch
Creating a branch from a base requires switching to the base branch
first, then creating the new branch and checking out to it, which
requires multiple operations.

Add base_branch parameter to create_branch to allow a new branch from a
base branch in one operation which is synonymous to the command "git
switch -c <new-branch> <base-branch>".

Signed-off-by: ayu-ch <ayu.chandekar@gmail.com>